Add expiremental 'compressed' messages module
Not standard json-rpc this is a 100% optional module that a
client can simply include to leverage compressed json-rpc
messages.

For a first draft, to compress messages we manually shorted
standard json-rpc header fields and compress / base64 encode
request / notification parmaters and response results.

Again this is expiremental and there is a good possibility
this will change going forward.